+++ tests/integration/ranking.test.ts
@@ -3,36 +3,39 @@ import request from 'supertest';
 
 import app from '../../src/app';
 import prisma from '../../src/config/prisma';
+import * as rankingService from '../../src/services/ranking.service';
+import { delCacheByPattern } from '../../src/utils/cache.util';
 
 describe('Ranking API', () => {
   const TEST_PREFIX = `TEST_RANKING_${Date.now()}`;
+  const BASE_TIME = 1000000000;
 
   let authToken: string;
   let myUserId: number;
 
   beforeEach(async () => {
-    // 추후 수정: totalTime 데이터는 향후 더 큰 값이나, db의 최댓값 + a로 값 수정이 필요합니다.
+    // ranking 정렬을 위해 기존 데이터보다 높은 값 사용
+    await delCacheByPattern('lampfire:ranking:*');
+
     const user1 = await prisma.user.create({
       data: {
         nickname: `${TEST_PREFIX}_USER_1`,
-        totalTime: 120000,
+        totalTime: BASE_TIME + 100,
       },
       select: { id: true },
     });
 
-    //const user2
     await prisma.user.create({
       data: {
         nickname: `${TEST_PREFIX}_USER_2`,
-        totalTime: 300000,
+        totalTime: BASE_TIME + 300,
       },
     });
 
-    // const user3
     await prisma.user.create({
       data: {
         nickname: `${TEST_PREFIX}_USER_3`,
-        totalTime: 200000,
+        totalTime: BASE_TIME + 200,
       },
     });
 
@@ -99,7 +102,6 @@ describe('Ranking API', () => {
   afterAll(async () => {
     await prisma.$disconnect();
   });
-
   // 전체 랭킹 조회
   describe('GET /rankings', () => {
     describe('200 - OK', () => {
@@ -111,62 +113,73 @@ describe('Ranking API', () => {
         expect(res.status).toBe(200);
         expect(res.body.success).toBe(true);
         expect(res.body.message).toBe('전체 랭킹');
-        // console.log('res.body,...', res.body);
-        expect(res.body.data).toHaveProperty('ranks');
-        expect(Array.isArray(res.body.data.ranks)).toBe(true);
 
-        expect(res.body.data.ranks.length).toBeGreaterThanOrEqual(3);
+        const { data } = res.body;
 
-        expect(res.body.data.ranks).toEqual(
+        expect(data).toHaveProperty('myRanking');
+        expect(data).toHaveProperty('topRankings');
+        expect(data).toHaveProperty('ranks');
+
+        expect(typeof data.myRanking.myRanking).toBe('number');
+        expect(Array.isArray(data.topRankings)).toBe(true);
+        expect(Array.isArray(data.ranks)).toBe(true);
+
+        expect(data.ranks.length).toBeGreaterThanOrEqual(3);
+
+        expect(data.ranks).toEqual(
           expect.arrayContaining([
             expect.objectContaining({
               nickname: `${TEST_PREFIX}_USER_1`,
-              totalTime: 120000,
+              totalTime: BASE_TIME + 100,
             }),
             expect.objectContaining({
               nickname: `${TEST_PREFIX}_USER_2`,
-              totalTime: 300000,
+              totalTime: BASE_TIME + 300,
             }),
             expect.objectContaining({
               nickname: `${TEST_PREFIX}_USER_3`,
-              totalTime: 200000,
+              totalTime: BASE_TIME + 200,
             }),
           ]),
         );
 
-        const rankingUsers = res.body.data.ranks.filter(
-          (user: { nickname: string }) => user.nickname.startsWith(TEST_PREFIX),
+        const rankingUsers = data.ranks.filter((user: { nickname: string }) =>
+          user.nickname.startsWith(TEST_PREFIX),
         );
 
         expect(rankingUsers[0]).toEqual(
           expect.objectContaining({
             nickname: `${TEST_PREFIX}_USER_2`,
-            totalTime: 300000,
+            totalTime: BASE_TIME + 300,
           }),
         );
 
         expect(rankingUsers[1]).toEqual(
           expect.objectContaining({
             nickname: `${TEST_PREFIX}_USER_3`,
-            totalTime: 200000,
+            totalTime: BASE_TIME + 200,
           }),
         );
 
         expect(rankingUsers[2]).toEqual(
           expect.objectContaining({
             nickname: `${TEST_PREFIX}_USER_1`,
-            totalTime: 120000,
+            totalTime: BASE_TIME + 100,
           }),
         );
+
+        if (rankingUsers.length >= 2) {
+          expect(rankingUsers[1].rank).toBeGreaterThan(rankingUsers[0].rank);
+        }
       });
     });
-
     // 401 - 잘못된 토큰일 경우, 인증 에러를 반환한다
     describe('401 - UNAUTHORIZED', () => {
       it('인증되지 않은 요청일 경우 반환한다', async () => {
         const invalidToken = jwt.sign({ id: 999999 }, process.env.JWT_SECRET!, {
           expiresIn: '10m',
         });
+
         const res = await request(app)
           .get('/rankings')
           .set('Cookie', [`token=${invalidToken}`]);
@@ -183,18 +196,18 @@ describe('Ranking API', () => {
         );
       });
     });
-
     // 500 - 서버 오류 시 명세된 에러 형식으로 반환한다
     describe('500 - INTERNAL_SERVER_ERROR', () => {
       it('서버 내부 예외가 발생할 경우 반환한다', async () => {
-        const prismaSpy = jest
-          .spyOn(prisma.user, 'findMany')
+        const serviceSpy = jest
+          .spyOn(rankingService, 'getMyRank')
           .mockRejectedValue(new Error('DB Error'));
 
         const res = await request(app)
           .get('/rankings')
           .set('Cookie', [`token=${authToken}`]);
-
+        // console.log('Status:', res.status);
+        // console.log('Body:', JSON.stringify(res.body, null, 2));
         expect(res.status).toBe(500);
         expect(res.body).toEqual(
           expect.objectContaining({
@@ -206,7 +219,7 @@ describe('Ranking API', () => {
           }),
         );
 
-        prismaSpy.mockRestore();
+        serviceSpy.mockRestore();
       });
     });
   });